refactor: formatted 이벤트 기간을 hook에서 넘겨받아 사용 가능하도록 수정해 재사용성 향상
