+++ packages/user/src/components/home/eventHero/LandingTitle.tsx
@@ -1,7 +1,9 @@
 import { PropsWithChildren } from 'react';
-import EventDurationText from './EventDurationText.tsx';
+import useGetEventDuration from 'src/hooks/query/useGetEventDuration.ts';
 
 export default function LandingTitle() {
+	const { formattedDuration } = useGetEventDuration();
+
 	return (
 		<Header>
 			<p className="text-body-4 font-medium">CASPER Electric 출시 이벤트</p>
@@ -13,9 +15,7 @@ export default function LandingTitle() {
 			</Subtitle>
 			<div className="flex items-center gap-3">
 				<Chip>이벤트 기간</Chip>
-				<p className="text-detail-1 h-[22px] w-[180px] min-w-max">
-					<EventDurationText />
-				</p>
+				<p className="text-detail-1 h-[22px] w-[180px] min-w-max">{formattedDuration}</p>
 			</div>
 		</Header>
 	);
